requirements.txt
python-dotenv==1.2.1
mediapipe==0.10.21
numpy==1.26.4
opencv-python==4.11.0.86
pyaudio==0.2.14
pyautogui==0.9.54
pynput==1.8.1
pygame==2.5.2
screeninfo==0.8.1
vosk==0.3.45